custom hook for likes, for reusebility
