+++ src/features/feed/components/Post.tsx
@@ -1,82 +1,19 @@
 "use client"
 import { Button } from '@/components/ui/button'
 import { formatDate } from '@/helpers/helper'
-import { likePost } from '@/services/post-interactions/post-interactions-service'
 import { Heart, MessageCircle, ThumbsUp } from 'lucide-react'
-import { getSession } from 'next-auth/react'
 import Link from 'next/link'
-import React, { useEffect, useState } from 'react'
 import { Skeleton } from '@/components/ui/skeleton'
 import { useSlugify } from '@/hooks/useSlugify'
-import { sendNotification } from '@/services/notifications/notification-service'
-import { NotificationType } from '@/features/notifications/types'
+import { useLikes } from '@/features/post/hooks/useLikes'
 
 // THIS IS ONLY FOR DEMO PURPOSES THIS CODE NEEDS TO BE REFACTORED LATTER FOR STABLE RELASE
 const Post = ({ post }: any) => {
-    const [isInteractionsLoading, setIsInteractionsLoading] = useState(true);
-    const [isLiked, setIsLiked] = useState<boolean | null>(null);
-    const [likes, setLikes] = useState<number | null>(null);
+    const { isLiked, isLikesLoading ,likes, handleLikePost} = useLikes(post);
     const { transformedSlug: slug } = useSlugify(post.author.username);
 
-    // **TODO** implement some rate limiting for likes
-    const handleLikePost = async (postId: string) => {
-        const user: any = await getSession();
-        
-        if(!user) return;
-
-        try {
-            // Optimistic UI update
-            setIsLiked(!isLiked);
-            setLikes((prev) => (prev ?? 0) + (isLiked ? -1 : 1));
-
-            // Call the API to like/unlike the post
-            await likePost(postId);
-
-            // Prevent self-like notifications
-            if(post.author.id === user?.user.id) return;
-
-            // Send notification only when post is liked
-            if(!isLiked){
-                await sendNotification(user?.user.id, user?.user.name, post.authorId, NotificationType.LIKE);
-            }
-
-        } catch (err) {
-            setIsLiked((prev) => !prev);
-            setLikes((prev) => (prev ?? 0) + (isLiked ? 1 : -1));
-            console.error("Error liking the post:", err);
-        }
-    };
-
-    useEffect(() => {
-        const fetchLikes = async () => {
-            try {
-                setIsInteractionsLoading(true);
-                const session: any = await getSession();
-                const userId = session?.user.id;
-
-                if (!userId) {
-                    setIsInteractionsLoading(false);
-                    setIsLiked(false);
-                    setLikes(post.Like.length);
-                    return;
-                }
-
-                const userLike = post.Like.find((like: any) => like.authorId === userId);
-                setIsLiked(!!userLike);
-                setLikes(post.Like.length);
-            } catch (err) {
-                console.error("Error fetching likes:", err);
-            }
-            finally{
-                setIsInteractionsLoading(false);
-            }
-        };
-
-        fetchLikes();
-    }, [post.Like]);
-
     return (
-            <Link href={`/post/${post.id}`} className="md:w-[600px] grid gap-2 h-auto bg-accent/50 hover:bg-accent/80 transition-all rounded-md p-3">
+            <div  className="md:w-[600px] w-full grid gap-2 h-auto bg-accent/50 hover:bg-accent/80 transition-all rounded-md p-3">
                 <div className="grid gap-1">
                     <div className="flex gap-2 w-full items-center">
                         <img className="h-8 w-8 rounded-full cursor-pointer" src={post.author.avatar as string | undefined} alt={post.author.username + ' ' + 'avatar'} />
@@ -85,11 +22,11 @@ const Post = ({ post }: any) => {
                     </div>
                 </div>
                 <div className='w-full'>
-                    <span className='font-bold'>{post.title}</span>
+                    <Link href={`/post/${post.id}`} className='font-bold hover:underline cursor-pointer transition-all'>{post.title}</Link>
                     <p className="line-clamp-4">{post.content}</p>
                 </div>
                 <div className='flex justify-start gap-2 items-center'>
-                    {!isInteractionsLoading ? (
+                    {!isLikesLoading ? (
                         <>
                             <Button onClick={() => handleLikePost(post.id)} className={`hover:text-primary cursor-pointer transition-all ${isLiked ? 'text-primary' : ''}`} variant={'secondary'}><ThumbsUp size={20} strokeWidth={0.75} />{likes}</Button>
                             <Button className={`hover:text-primary cursor-pointer transition-all`} variant={'secondary'}><MessageCircle size={20} strokeWidth={0.75} /></Button>
@@ -103,7 +40,7 @@ const Post = ({ post }: any) => {
                         </>
                     )}
                 </div>
-            </Link>
+            </div>
         )
     }
